requirements.txt
numpy 
pandas 
seaborn
matplotlib
apache-airflow
apache-airflow[auth]
scikit-learn



